# docs/openapi.json
{
  "openapi": "3.0.0",
  "paths": {
...
    "/api/organizations": {
      "post": {
        "operationId": "OrganizationsController_create",
        "parameters": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CreateOrganizationDto"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": ""
          }
        },
        "security": [
          {
            "JWT-auth": []
          }
        ],
        "tags": [
          "Organizations"
        ]
      },
      "get": {
        "operationId": "OrganizationsController_findAll",
        "parameters": [],
        "responses": {
          "200": {
            "description": ""
          }
        },
        "security": [
          {
            "JWT-auth": []
          }
        ],
        "tags": [
          "Organizations"
        ]
      }
    },
    "/api/organizations/{id}": {
      "get": {
        "operationId": "OrganizationsController_findOne",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": ""
          }
        },
        "security": [
          {
            "JWT-auth": []
          }
        ],
        "tags": [
          "Organizations"
        ]
      },
      "patch": {
        "operationId": "OrganizationsController_update",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/UpdateOrganizationDto"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": ""
          }
        },
        "security": [
          {
            "JWT-auth": []
          }
        ],
        "tags": [
          "Organizations"
        ]
      },
      "delete": {
        "operationId": "OrganizationsController_remove",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "204": {
            "description": ""
          }
        },
        "security": [
          {
            "JWT-auth": []
          }
        ],
        "tags": [
          "Organizations"
        ]
      }
    },
...
    "/api/organizations/{organizationId}/credits": {
      "get": {
        "operationId": "CreditsController_getBalance",
        "parameters": [],
        "responses": {
          "200": {
            "description": ""
          }
        },
        "security": [
          {
            "JWT-auth": []
          }
        ],
        "tags": [
          "Credits"
        ]
      }
    },
...
    "/api/organizations/{organizationId}/invoices": {
      "get": {
        "operationId": "InvoicesController_findAll",
        "parameters": [
          {
            "name": "status",
            "required": false,
            "in": "query",
            "description": "Filter nach Rechnungsstatus",
            "schema": {
              "example": "paid",
              "type": "string",
              "enum": [
                "draft",
                "pending",
                "paid",
                "cancelled"
              ]
            }
          },
          {
            "name": "startDate",
            "required": false,
            "in": "query",
            "description": "Startdatum für Filterung",
            "schema": {
              "example": "2024-01-01T00:00:00.000Z",
              "type": "string"
            }
          },
          {
            "name": "endDate",
            "required": false,
            "in": "query",
            "description": "Enddatum für Filterung",
            "schema": {
              "example": "2024-12-31T23:59:59.000Z",
              "type": "string"
            }
          },
          {
            "name": "page",
            "required": false,
            "in": "query",
            "description": "Seitennummer für Pagination",
            "schema": {
              "default": 1,
              "example": 1,
              "type": "number"
            }
          },
          {
            "name": "limit",
            "required": false,
            "in": "query",
            "description": "Anzahl der Ergebnisse pro Seite",
            "schema": {
              "default": 20,
              "example": 20,
              "type": "number"
            }
          }
        ],
        "responses": {
          "200": {
            "description": ""
          }
        },
        "security": [
          {
            "JWT-auth": []
          }
        ],
        "tags": [
          "Invoices"
        ]
      }
    },
...
    "/api/admin/organizations": {
      "get": {
        "operationId": "AdminController_findAllOrganizations",
        "parameters": [
          {
            "name": "search",
            "required": false,
            "in": "query",
            "description": "Suchbegriff für Organisationsname",
            "schema": {
              "example": "Musterverein",
              "type": "string"
            }
          },
          {
            "name": "page",
            "required": false,
            "in": "query",
            "description": "Seitennummer für Pagination",
            "schema": {
              "default": 1,
              "example": 1,
              "type": "number"
            }
          },
          {
            "name": "limit",
            "required": false,
            "in": "query",
            "description": "Anzahl der Ergebnisse pro Seite",
            "schema": {
              "default": 20,
              "example": 20,
              "type": "number"
            }
          }
        ],
        "responses": {
          "200": {
            "description": ""
          }
        },
        "security": [
          {
            "JWT-auth": []
          }
        ],
        "tags": [
          "Admin"
        ]
      }
    },
    "/api/admin/organizations/{id}": {
      "get": {
        "operationId": "AdminController_getOrganization",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": ""
          }
        },
        "security": [
          {
            "JWT-auth": []
          }
        ],
        "tags": [
          "Admin"
        ]
      },
      "patch": {
        "operationId": "AdminController_updateOrganization",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/UpdateOrganizationAdminDto"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": ""
          }
        },
        "security": [
          {
            "JWT-auth": []
          }
        ],
        "tags": [
          "Admin"
        ]
      }
    },
    "/api/admin/organizations/{id}/credits": {
      "patch": {
        "operationId": "AdminController_adjustCredits",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/AdjustCreditsDto"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": ""
          }
        },
        "security": [
          {
            "JWT-auth": []
          }
        ],
        "tags": [
          "Admin"
        ]
      }
    },
    "/api/admin/organizations/{id}/discount": {
      "patch": {
        "operationId": "AdminController_setDiscount",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/SetDiscountDto"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": ""
          }
        },
        "security": [
          {
            "JWT-auth": []
          }
        ],
        "tags": [
          "Admin"
        ]
      },
      "delete": {
        "operationId": "AdminController_removeDiscount",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": ""
          }
        },
        "security": [
          {
            "JWT-auth": []
          }
        ],
        "tags": [
          "Admin"
        ]
      }
    },
    "/api/admin/organizations/{id}/access": {
      "post": {
        "operationId": "AdminController_accessWithPin",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/AccessOrganizationDto"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": ""
          }
        },
        "security": [
          {
            "JWT-auth": []
          }
        ],
        "tags": [
          "Admin"
        ]
      }
    },
    "/api/admin/organizations/{id}/impersonate": {
      "get": {
        "operationId": "AdminController_impersonate",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": ""
          }
        },
        "security": [
          {
            "JWT-auth": []
          }
        ],
        "tags": [
          "Admin"
        ]
      }
    },
    "/api/admin/users": {
      "get": {
        "operationId": "AdminController_findAllUsers",
        "parameters": [
          {
            "name": "search",
            "required": false,
            "in": "query",
            "description": "Suchbegriff für Benutzername oder E-Mail",
            "schema": {
              "example": "max@example.com",
              "type": "string"
            }
          },
          {
            "name": "isLocked",
            "required": false,
            "in": "query",
            "description": "Filter nach gesperrten Benutzern",
            "schema": {
              "example": false,
              "type": "boolean"
            }
          },
          {
            "name": "page",
            "required": false,
            "in": "query",
            "description": "Seitennummer für Pagination",
            "schema": {
              "default": 1,
              "example": 1,
              "type": "number"
            }
          },
          {
            "name": "limit",
            "required": false,
            "in": "query",
            "description": "Anzahl der Ergebnisse pro Seite",
            "schema": {
              "default": 20,
              "example": 20,
              "type": "number"
            }
          }
        ],
        "responses": {
          "200": {
            "description": ""
          }
        },
        "security": [
          {
            "JWT-auth": []
          }
        ],
        "tags": [
          "Admin"
        ]
      }
    },
    "/api/admin/users/{id}/unlock": {
      "post": {
        "operationId": "AdminController_unlockUser",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "201": {
            "description": ""
          }
        },
        "security": [
          {
            "JWT-auth": []
          }
        ],
        "tags": [
          "Admin"
        ]
      }
    },
    "/api/admin/purchases": {
      "get": {
        "operationId": "AdminController_findAllPurchases",
        "parameters": [
          {
            "name": "organizationId",
            "required": false,
            "in": "query",
            "description": "Filter nach Organisations-ID",
            "schema": {
              "example": "550e8400-e29b-41d4-a716-446655440000",
              "type": "string"
            }
          },
          {
            "name": "status",
            "required": false,
            "in": "query",
            "description": "Filter nach Zahlungsstatus",
            "schema": {
              "example": "completed",
              "type": "string",
              "enum": [
                "pending",
                "completed",
                "failed",
                "refunded"
              ]
            }
          },
          {
            "name": "startDate",
            "required": false,
            "in": "query",
            "description": "Startdatum für Filterung",
            "schema": {
              "example": "2024-01-01T00:00:00.000Z",
              "type": "string"
            }
          },
          {
            "name": "endDate",
            "required": false,
            "in": "query",
            "description": "Enddatum für Filterung",
            "schema": {
              "example": "2024-12-31T23:59:59.000Z",
              "type": "string"
            }
          },
          {
            "name": "page",
            "required": false,
            "in": "query",
            "description": "Seitennummer für Pagination",
            "schema": {
              "default": 1,
              "example": 1,
              "type": "number"
            }
          },
          {
            "name": "limit",
            "required": false,
            "in": "query",
            "description": "Anzahl der Ergebnisse pro Seite",
            "schema": {
              "default": 20,
              "example": 20,
              "type": "number"
            }
          }
        ],
        "responses": {
          "200": {
            "description": ""
          }
        },
        "security": [
          {
            "JWT-auth": []
          }
        ],
        "tags": [
          "Admin"
        ]
      }
    },
    "/api/admin/purchases/{id}/complete": {
      "post": {
        "operationId": "AdminController_completePurchase",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "201": {
            "description": ""
          }
        },
        "security": [
          {
            "JWT-auth": []
          }
        ],
        "tags": [
          "Admin"
        ]
      }
    },
    "/api/admin/invoices": {
      "get": {
        "operationId": "AdminController_findAllInvoices",
        "parameters": [
          {
            "name": "organizationId",
            "required": false,
            "in": "query",
            "description": "Filter nach Organisations-ID",
            "schema": {
              "example": "550e8400-e29b-41d4-a716-446655440000",
              "type": "string"
            }
          },
          {
            "name": "status",
            "required": false,
            "in": "query",
            "description": "Filter nach Rechnungsstatus",
            "schema": {
              "example": "paid",
              "type": "string",
              "enum": [
                "draft",
                "pending",
                "paid",
                "cancelled"
              ]
            }
          },
          {
            "name": "startDate",
            "required": false,
            "in": "query",
            "description": "Startdatum für Filterung",
            "schema": {
              "example": "2024-01-01T00:00:00.000Z",
              "type": "string"
            }
          },
          {
            "name": "endDate",
            "required": false,
            "in": "query",
            "description": "Enddatum für Filterung",
            "schema": {
              "example": "2024-12-31T23:59:59.000Z",
              "type": "string"
            }
          },
          {
            "name": "page",
            "required": false,
            "in": "query",
            "description": "Seitennummer für Pagination",
            "schema": {
              "default": 1,
              "example": 1,
              "type": "number"
            }
          },
          {
            "name": "limit",
            "required": false,
            "in": "query",
            "description": "Anzahl der Ergebnisse pro Seite",
            "schema": {
              "default": 20,
              "example": 20,
              "type": "number"
            }
          }
        ],
        "responses": {
          "200": {
            "description": ""
          }
        },
        "security": [
          {
            "JWT-auth": []
          }
        ],
        "tags": [
          "Admin"
        ]
      }
    },
    "/api/admin/invoices/{id}/mark-paid": {
      "post": {
        "operationId": "AdminController_markInvoicePaid",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "201": {
            "description": ""
          }
        },
        "security": [
          {
            "JWT-auth": []
          }
        ],
        "tags": [
          "Admin"
        ]
      }
    },
...
    "/api/admin/rental-hardware/{id}": {
      "patch": {
        "operationId": "AdminController_updateRentalHardware",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/UpdateRentalHardwareDto"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": ""
          }
        },
        "security": [
          {
            "JWT-auth": []
          }
        ],
        "tags": [
          "Admin"
        ]
      },
      "delete": {
        "operationId": "AdminController_deleteRentalHardware",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": ""
          }
        },
        "security": [
          {
            "JWT-auth": []
          }
        ],
        "tags": [
          "Admin"
        ]
      }
    },
...
    "/api/admin/rental-assignments/{id}/return": {
      "post": {
        "operationId": "AdminController_returnRental",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "201": {
            "description": ""
          }
        },
        "security": [
          {
            "JWT-auth": []
          }
        ],
        "tags": [
          "Admin"
        ]
      }
    },
    "/api/admin/stats/overview": {
      "get": {
        "operationId": "AdminController_getOverviewStats",
        "parameters": [],
        "responses": {
          "200": {
            "description": ""
          }
        },
        "security": [
          {
            "JWT-auth": []
          }
        ],
        "tags": [
          "Admin"
        ]
      }
    },
...
  },
  "info": {
    "title": "OpenEOS API",
    "description": "\n# OpenEOS API Documentation\n\nOpen Event Ordering System - REST API für Kassensysteme bei Vereinen und Veranstaltungen.\n\n## Authentifizierung\n\nDie API verwendet JWT Bearer Tokens für die Authentifizierung. Nach dem Login erhältst du einen Access Token (30 Min gültig) und einen Refresh Token (7 Tage, als httpOnly Cookie).\n\n### Endpoints ohne Authentifizierung\n- `POST /api/auth/register` - Registrierung\n- `POST /api/auth/login` - Login\n- `POST /api/auth/refresh` - Token erneuern\n- `POST /api/auth/forgot-password` - Passwort vergessen\n- `POST /api/auth/reset-password` - Passwort zurücksetzen\n- `GET /api/health/*` - Health Checks\n- `GET /api/invitations/:token` - Einladung abrufen\n- `/api/public/order/*` - Online-Bestellungen (mit Session-Token)\n\n## Multi-Tenant\n\nDie API ist multi-tenant-fähig. Die meisten Endpoints erfordern eine `organizationId` im Pfad.\n\n## Rollen\n\n- **ADMIN** - Vollzugriff auf Organisation\n- **MANAGER** - Verwaltung von Produkten, Events, Bestellungen\n- **CASHIER** - Kassen-Operationen\n- **KITCHEN** - Küchen-Display\n- **DELIVERY** - Ausgabe-Display\n\n## WebSocket\n\nReal-time Updates über Socket.io unter `/` mit Namespaces:\n- `/pos` - Kassen-Clients\n- `/display` - Küchen/Ausgabe-Displays\n- `/printer` - Drucker-Agenten\n- `/admin` - Admin-Dashboard\n    ",
    "version": "1.0.0",
    "contact": {
      "name": "OpenEOS",
...
    },
    "license": {
...
  "servers": [
...
  "components": {
    "securitySchemes": {
      "JWT-auth": {
        "scheme": "bearer",
        "bearerFormat": "JWT",
        "type": "http",
        "name": "JWT",
        "description": "Enter JWT token",
        "in": "header"
      },
      "X-Organization-Id": {
        "type": "apiKey",
        "in": "header",
        "name": "X-Organization-Id",
        "description": "Organization ID for multi-tenant requests"
      },
      "X-Session-Token": {
        "type": "apiKey",
        "in": "header",
        "name": "X-Session-Token",
        "description": "Session token for online orders"
      }
    },
...
  }
}